// src/openapi/models/count-metric-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * TRON Common API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface CountMetricDto
 */
export interface CountMetricDto {
    /**
     * 
     * @type {string}
     * @memberof CountMetricDto
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof CountMetricDto
     */
    path?: string;
    /**
     * 
     * @type {number}
     * @memberof CountMetricDto
     */
    sum?: number;
}